requirements.txt
openai==1.58.1
python-dotenv
streamlit == 1.49.1